move: ProductPriceChangeEvent 클래스를 shared/event로 이동
